Add test configuration for database access
This package must work with either psycopg2 or psycopg(3).

This means that some setup is required given that these two library
versions are not 100% compatible. More specifically, the name of some
classes and their import paths and arguments are different.

In order to create an abstraction so that the code in this package can
deal with either versions without too much friction, a
src/psycopack/_psycopg.py file is added to contain the necessary logic
pertaining to psycopg imports so that this issue is out of the way.

A pytest fixture that will form the basis of the db connection and setup
is added. It also provides evidence this commit works for either
version of psycopg (test_fixtures.py).

Some options of 3rd-party packages were considered to handle the pytest
setup for working with psycopg, but they either don't support psycopg2
[1], which is a deal breaker or they haven't seen an update in a long
while [2]. Also they provide more functionality (and less control) than
what may be needed.

[1] - https://github.com/dbfixtures/pytest-postgresql
[2] - https://github.com/CloverHealth/pytest-pgsql