Add Repack class
This commit releases the first interaction of Repack, which includes a
function Repack(...).full() to process a table repack from beginning to
end.

There are several limitations in this first take, mostly around the
underlying functionality not being re-entrant and idempotent. These
limitations will be worked out in subsequent pull requests, as they are
must-have features before this functionality can be used in production.

At present, the idea is to showcase how this class can be used, along
with the basic interface that allows for customisation and control over
each step of the repacking process.

This "control over the process" is the primary difference between this
package and the pg_repack extension. The Repack(...).full() achieves the
same result as pg_repack, and is, in a way, redundant - but serves the
purpose of displaying what an end-to-end repacking requires.