Deduplicate repack assertion logic
This commit extracts the assertion logic so that it can be reused by
future tests without relying on copy/paste.

Context: To verify that the repacking process has finished successfully
we have to:

1. Verify that the repacked table has a different oid. Meaning it's a
   different table.
2. Verify that all indexes, constraints, and referring foreign keys are
   the same as the old table.
3. Verify that the triggers and functions created as part of the
   repacking process have been removed.

Given that these assertions will be universal giving the repacking
approach used in this project, they will be required for every
full-repacking test going forward.