Add wrapper to execute SQL statements with logging
The new class "LoggedCursor" will be used to execute SQL statements.
This class is a thin wrapper around Pyscopg's Cursor, with the subtle
difference that it logs queries and errors using Python's logging
module.

For example, running the following queries through the executor:

  SELECT 1;
  SELECT * FROM table_that_does_not_exist;

Produces the following log outputs:

  DEBUG:psycopack:SELECT 1;
  DEBUG:psycopack:SELECT * FROM table_that_does_not_exist;
  ERROR:psycopack:Failed to execute statement: SELECT * FROM table_that_does_not_exist;
  Traceback (most recent call last):
    ... summarised ...
    File "/.../site-packages/psycopg/cursor.py", line 97, in execute
      raise ex.with_traceback(None)
  psycopg.errors.UndefinedTable: relation "table_that_does_not_exist" does not exist
  LINE 1: SELECT * FROM table_that_does_not_exist;

Context: Prior to this commit, there was no way to see what queries the
repacking process has triggered, unless the user had access to Postgres
logfile (which may not be available in certain installations). Ideally,
all queries should be emitted to stdout so that the user could both
validate the queries and learn about the process.

In a subsequent commit, the existing Cursor calls will be swapped by
LoggedCursor calls.