Add missing .as_string() call
These three statements where missing an as_string() call. That means
that self.cur.execute() was receiving a Composed object instead of a
string.

While this is not a big issue, subsequent commits will add and use a
wrapper around the Cursor execute() function so that sql statements can
be logged using Python's logging module. The logger will in turn require
a string to be called with logger.debug(), logger.error(), etc.

As a side effect, this also aids the overall consistency, as the
majority of queries use as_string().