Introduce a repacking tracker
Prior to this commit, there was no way to recover a repacking process
from failure or command interruption.

Under those circumstances, the repacking process would need to start
from scratch and the existing repacking would be wasted.

This can be costly, as repacking a whole table can take a lot of time.

This commit introduces a "Tracker" class. Its goal is to keep track of
the repacking process in its totality, so that if it is interrupted, it
can resume from where it left off instead of having to run from scratch.

From a user's perspective, the only thing they need to do call
Repack.full() until the process finishes. The same as before.

The tracker class acknowledges all the existing repacking stages, and
provides verification to ensure the caller that the current stage of the
repacking hasn't been messed up with.

This is relevant as a user might manually (or accidentally) delete
relations created and required by the repacking process while the
process is still running. Although in many cases this is not a
salvageable situation, the tracker tries its best to return a meaningful
error.

The way the tracker provides idempotency/reentrancy to the repacking
process is by storing information about the completed and ongoing
repacking stages in an auxiliary table, and using this table to verify
the current integrity of the process.