Add revert_swap operation
Prior to this commit, the swap() operation was not revertible
(unrecoverable from).

The problem is that once the copy table (repacked) had been swapped with
the original, there was no way to go back on it.

Any potential problems that may have been caused by the swap operation
would therefore be permanent.

This change introduces the revert_swap() function, which will allow for
a way to recover from the swap() operation (should that be necessary).

This new operation restores the original state of both original and copy
tables before they were swapped.

After the revert_swap() operation is called, the swap() operation can be
called again.