Change relation prefixes to be more identifiable
Prior to this change, the relations involved in the repacking process
were prefixed by the name "repack".

One potential issue is that a particular user looking at the database
schema while the repacking process is happening might be confused about
why those relations exist and what they mean.

This change alters the prefix name to "psycopack". This way, with a more
unique prefix, the unaware user can search for it and find the reason
for these relations to exist.

As an example, the relation names before a swap for table oid 12345
will be:

 psycopack_12345                   (copy table)
 psycopack_12345_backfill          (backfill log table)
 psycopack_12345_backfill_id_seq   (id seq of the above)
 psycopack_12345_tracker           (repacking tracker table)
 psycopack_12345_id_seq            (id seq of the above)

And after the swap, but before the clean up:

 psycopack_12345_backfill          (backfill log table)
 psycopack_12345_backfill_id_seq   (id seq of the above)
 psycopack_12345_tracker           (repacking tracker table)
 psycopack_12345_id_seq            (id seq of the above)
 psycopack_repacked_12345          (original table, now repacked)