Generalise get_index_def function
Prior to this change, we had two functions for fetching index definition
from Postgres.

- get_index_def
- get_non_pk_index_def

The latter was merely a shadow of the former, with a little hack to
remove indexes corresponding to the primary key.

This commit generalises the get_index_def by incorporating a new boolean
into the data class definition to host information on whether the index
is backing up a primary key [1].

This work will also facilitating pulling more information from the
pg_index table in future when psycopack starts to handle exclusion
constraints.

[1] https://www.postgresql.org/docs/current/catalog-pg-index.html