Handle tables with exclusion constraints
Prior to this change, there was no support for tables with exclusion
constraints.

This commit changes that, by updating the setup stage to include the
exclusion constraint before the table is backfilled.

As a side-effect, tables with exclusion constraints will have slower
backfills. This is because the index that backs up the exclusion
constraint will have to be looked up for each insert, slowing writes to
the table.

There is no workaround for this, however. Postgres does not support a
"CREATE INDEX CONCURRENTLY" followed up by an ALTER TABLE ... ADD
CONSTRAINT ... USING <that index>.

This means that we can't create exclusion constraints ONLINE, and
therefore the best place to add an exclusion constraint is at the
beginning, when the copy table is still empty.