Add check against inherited tables
Prior to this commit, inherited tables were not being checked against.

We don't have a use case for handling inherited tables at the present,
and as such, it does not make sense to support them for the first relase
of psycopack.

In future, this decision might change and we might be interested in
supporting inherited tables.