Fix validation error when multiple indexes present
Prior to this change, the algorithm to rename indexes (from the copy
table to the original) did not work when the original table had
duplicated indexes (indexes on the same field with the same definition).

The reason, is that the algorithm uses a Python dictionary where the
keys are a stripped-down version of the index definition. That means
that a second index with the same definition would override the existing
key in that dictionary, instead of appending to it.

This change fixes that, by using a list of indexes that have the
particular index definition of the dictionary key, and iterating through
it when renaming indexes.