Assert table must have supported primary key
Prior to this change, there was nothing preventing a table with
unsupported primary keys (or lack thereof), to run through the psycopack
process. The process would fail unexpectedly in the table setup stage,
however.

This change makes sure that tables with unsupported primary keys do not
even start the psycopack process.

The check covers:

1. Tables without a PK.
2. Tables with a composite PK.
3. Tables where the PK column is not named "id".
4. Tables where the PK type is not supported (int or serial and their
   variants).