Add support to convert pk to bigint
When repacking the original table, it might be handy to also upgrade the
primary key field to bigint.

This change introduces a new flag "convert_pk_to_bigint" that will
change the pk column type during the setup (when the copy table is
created).

NOTE: Whilst identity pks and pks with manual sequences are handled,
psycopack does not yet handle SERIAL pks explicitly - this will need
tests and handling in subsequent PRs.