Handle renaming of primary key sequences
Prior to this change, the primary key sequence name was not being
swapped along.

This meant that the newly repacked table would remain with a sequence
name like: "psycopack_12345_id_seq" instead of something like
"foo_id_seq", where "foo" is the original table name.

This commit changes that by ensuring that at the end of the repacking
process, the new table will have the correct sequence name.