Add test for table with serial primary key
Prior to this change there was no explicit test for verifying the
repacking process for a table with a serial PK.

These types of table are just like tables with manually-set primary key
default sequences, and so, they work in the same way and are repacked in
the same way. This test confirms it.