Generalise pk column name
Prior to this change, the code enforced a hardcoded requirement for the
primary key column to be named "id".

This limitation is unecessary. The primary key column can have any
arbitrary name.

This change introduces an instance variable (self.pk_column) for the
primary key, which is determined through introspection queries.

This change allows for all the introspection queries and commands that
have the hardcoded requirement above to be updated to use the instance
variable instead.