Make backfill table handle bigint values
Prior to this change, the backfill table looked like this:

  CREATE TABLE ... (
    id SERIAL PRIMARY KEY,
    batch_start INT,
    batch_end INT,
    finished BOOLEAN

That meant that tables with large integer values (greater than a
unsigined 32bit integer), wouldn't fit in the table.

Local experimentation proved that psycopg would raise the following
exception when this situation was attemped:

  E           psycopg.errors.NumericValueOutOfRange: integer out of range

This change fixes this problem, by increasing the batch_start and
batch_end data types to be big integers.

The signature for the repacking function also had to be updated so that
the batch inserts (which call that function) wouldn't fail.