Handle tables without pk sequence
Prior to this change, the psycopack process was assuming the primary key
would have a sequence to back it up.

This is not necessarily true, as the user might decide to insert ids
manually into the table and keep track of them in any arbitrary way.

This may not be a realistic scenario for most tables that will be
processed by psycopack, but it's still worth handling.

This change skips the parts of the code that where engaging with primary
key sequences.