Add post-backfill batch callback
Prior to this change, once the psycopack process started backfilling a
table there was no feedback mechanism for the caller to get information
about how the backfill was performing.

This change introduces a new argument (post_backfill_batch_callback) to
the Repack class.

This new argument allows a caller to pass a callback function that will
be called each time the backfill for a batch is finished.

This allows for better instrumentation. For example, the caller might
want to programmatically push metrics on how the progress of the
Repack backfill is going on their production environment so that they
can visualise this data in a dashboard of their choice.