Make clean_up routine happen inside transaction
Prior to this commit, if the clean_up routine failed half-way through,
the new table might end up with the wrong index or constraint names.

This change makes the DDLs run inside a transaction. This makes it
easier to control for idempotency.

The statements in this method are fairly fast as they only update
catalogue tables so the accumulation of locks shouldn't block the
database for too long.

Note that the introspection queries are all executed before the code
starts the transaction to make the transaction length as short as
possible.