Simplify referring FK clean up logic
Prior to this change, the logic to rename the referring foreign key
constraints was something like that (mock code):

  define swap (x, y):
      temp := x
      x := y
      y := temp

Only that, in the lines succeeding, we drop the foreign key constraint
from the referring table.

This change thus simplify the logic, by deleting the referring foreign
key constraint first, so that there is no name conflict when the shadow
table constraint is renamed.

This will avoid having to do the swap with a temporary name, and also
means we'll avoid one ALTER TABLE command per referring foreign key,
which will make the clean-up process faster.