Change instruction to create indexes
Prior to this change, if the schema sync stage failed midway through,
it would have potentially left some indexes behind.

When the user runs the sync stage again to finish the process, it would
get stuck in the index creation because the index of the same name
already exists.

This change adds the IF NOT EXISTS construct so that the schema sync
stage can skip index creation when the index already exists.

This work goes towards making the schema sync stage idempotent and
reentrant and so that test_sync_schemas_is_reentrant_and_idempotent can
stop being skipped.