Do not try to recreate existing unique constraints
Prior to this change, if the process of creating unique constraints
failed midway through, some unique constraints would be created and some
would not.

This change introduces a new logic to check if the constraint is already
created before trying to create it again.

This is part of the solution towards making the schema sync stage
idempotent and reentrant.