Do not try to recreate CHECK and FK constraints
Prior to this change, if the process of creating CHECK and FK
constraints failed midway through, some constraints would be created and
some would not. They also might have ended up being left as NOT VALID
when they should not.

This change introduces a new logic to check if the constraint is already
created before trying to create it again. It also checks if the
validation is as expected and act accordingly if not.

This is part of the solution towards making the schema sync stage
idempotent and reentrant.