Do not try to recreate referring FK constraints
Prior to this change, if the process of creating a referring FK [0]
failed midway through, some referring fks would be created and some
would not. They also might have ended up being left as NOT VALID when
they should not.

This change introduces a new logic to check if the constraint is already
created before trying to create it again. It also checks if the
validation is as expected and acts accordingly if not.

This is the final part of the solution towards making the schema sync
stage idempotent and reentrant.

[0] A referring FK of table A is an FK on table B that points to table
    A.