Add test to cover sync_schema idempotency
The current code does not guarantee idempotency when the sync_schema
function is called.

This commit merely adds a test that will cover idempotency when the next
commits fix this problem.

The test is initially skipped as to not break CI, but should be enabled
once all changes are consolidated.