Introduce lock timeouts
Prior to this change, psycopack had no consideration for lock timeouts.

This means that if the underlying table being psycopacked was busy, the
lock queue would be haulted when psycopack AT commands tried to acquire
an access exclusive lock on the table. This would potentially result in
a outage.

This change introduces a configurable lock timeout parameter, so that
the user can control the behaviour of how long to wait to acquire a lock
before bailing out.

This also introduces room for more failures and interruptions during the
process. But considering that each stage is idempotent and reentrant,
the user should be able to run the same stage again during less-busy
hours.

The timeout strategy for each stage is as follows:

- pre_setup: Doesn't need lock timeouts as it only fires introspection
  queries.
- setup: Creates the copy relations, the trigger, and the function. The
  whole process is behind the lock_timeout ctx manager.
- backfill: Only performs reads and writes, so there's no use for lock
  timeouts here
- sync_schemas: The index creation is particular, and requires
  lock_timeouts set to 0, apart from that, the other functions should
  all obey the set lock_timeout value.
- swap/revert_swap/clean_up: since these operations are all done within
  a transaction, the lock_timeout value is applied on it.