Do not accept table with invalid indexes
Prior to this change, Psycopack did not discriminate against tables with
invalid indexes.

The problem is that this situation is ambiguous. Does the user know the
index is invalid? Do they want the index to be fixed in the new table?
Or should it be dropped?

Instead of making an arbitrary decision, this change introduces a check
that will bailout during pre-validation should the table have any
invalid index.