Recreate invalid indexes when syncing schemas
Prior to this change, the sync_schemas didn't deal with existing invalid
indexes in the copy table.

Invalid indexes could've been caused by interruptions during the
sync_schemas routine.

If the interruption happened during the index creation, it is possible
that the index would be left in an invalid state.

This commit changes the underlying logic for creating the indexes in the
copy table in a way that existing invalid indexes are dropped
beforehand, so that they can be recreated by the next steps.