Add introspector and command to Tracker class
Prior to this change, the Tracker class was instantiating its own
command and introspector class.

In future, these two classes will need to be schema-aware, and thus
receive a `schema` argument.

In order to avoid passing the same variable down the chain multiple
times, this change will make so that the Tracker class takes in the
already-instantiate introspector and command instances.