Make Introspection schema aware
Prior to this change, there was a hard assumption in the Introspector
class that Psycopack would be running in the "public" schema. This might
not be the case.

This change introduces a "schema" argument to Repack that is passed down
to the Introspector class.

This change is the first part (out of three), to make the whole
Pyscopack process schema aware. Subsequent commits address the second
and thirdpart, involving making the Command (DDL) and the tracker
statements schema aware.