Make Command schema aware
Prior to this change, there was a hard assumption in the Command class
that Psycopack would be running in the "public" schema. This might not
be the case.

This change passes down the "schema" argument from Repack to the Command
class.

This change is the second part (two out of three), to make the whole
Pyscopack process schema aware.