Make Tracker schema aware
Prior to this change, there was a hard assumption in the Tracker class
that Psycopack would be running in the "public" schema. This might not
be the case.

This change passes down the "schema" argument from Repack to the Tracker
class.

This change is the last part (three out of three), to make the whole
Pyscopack process schema aware.

A test has also been added to ensure all repacking steps work when the
table is in a different schema.