Forbid repacking when FK is in another schema
Prior to this change, Psycopack would error halfway through because it
doesn't know how to handle referring foreign keys from another schema.

Postgres allows a table from one schema to refer to another in another
schema. This allows for Postgres users who are interested in having
modelling such as a "users" schema with a user table to be linked to an
"accounts" schema with an account table.

Some adjustments are needed to Psycopack to make this work. But for now,
this change just raises an exception before starting the process.

Later on this can be adjusted to properly handle such situations. Right
now it is not part of our original usecase.