User must be OWNER of table to be Psycopacked
Prior to this change, there was no check to verify that the user was an
OWNER of the table to be repacked.

This is a problem because only the owner can rename and delete the table
(needed once it has been swapped).

This change adds an explicit check to ensure the user is the owner of
the table before proceeding.