User must be OWNER of referring tables
Prior to this change, there was no check to verify that the user was an
OWNER of the referring tables (tables that have fks pointing to the
table to be repacked).

This is a problem because Psycopack needs to create new foreign keys on
that table that point to the shadow table, and that is not possible
unless the user is the table owner.

This change adds an explicit check to ensure the user is the owner of
the referring tables before proceeding.