User must have REFERENCES privilege on referred tables
Prior to this change, there was no check to verify that the user had the
REFERENCES privilege on the referred tables (tables the table to be
repacked has a fk to).

This is a problem because Psycopack needs to create new foreign keys on
the shadow table that point to the referred tables, and that is not
possible unless the user has REFERENCES privilege.

This change adds an explicit check to ensure the user has this privilege
for all referred tables before proceeding.