Check table exists and user permissions early
Prior to this change, checking if the table existed was a side-effect of
finding the name of the copy table (which requires querying the original
table OID).

Additionally, the check for user permissions happened as the last logic
in initialisation.

This change moves those checks up so that the Repack class can bailout
early.