Introduce the Registry
Prior to this change, there was no "global state" stored away that the
Psycopack process could rely on in case it failed and had to start
again.

This meant that subtle bugs such as reinstantiating the Psycopack class
after performing a swap operation would culminate in failure as
Psycopack couldn't determine the proper state of the repacking process
and the name of its objects.

This change introduces a Registry, which is responsible for keeping the
name of objects relating to the Psycopack process of a table, which will
fix the bug mentioned above.

The Registry also has a nice side-effect: it can be used to find all the
name of objects involved in the Psycopack process. Previously, one of
the ways of doing that was via `\d *psycopack*`, which would return a
bunch of noise (indexes, and such). But now the Registry table can be
queried directly instead.

This change will also make it more manageable when the user using
Psycopack in many tables, as they can refer to the registry.