Rename Repack class to Psycopack
Prior to this change, this class was named Repack after the pg_repack
Postgres extension, as it served as inspiration for this code.

However, in future the documentation will refer to this class as the
entry point for the Python interface of the package.

Renaming the Repack class to Psycopack will help newbie users (or
experienced users who don't know what "repack" means in the context of
Postgres) to not have to load more context into their brain if all they
want to do is process a schema change safely.