Add function to fetch a cursor
Prior to this change, this package was relying on the default behaviour
of Psycopg for cursor instantiation.

The problem is that there are differences between Psycopg 2 and Psycopg
3 that make this unfeasible for this library, which supports both
versions.

For example, when a cursor is instantiated from the connection context
manager, Psycopg 2 starts a transaction automatically [1][2], whereas
Psycopg 3 doesn't.

  with connection.cursor() as cur:
      # There is already an implicit "BEGIN;" here on Psycopg 2.
      # But not on Psycopg 3
      ...

This problem came up when testing Psycopack locally using Psycopg 2.

The following ABORT call had to be added or else the procedure would
fail when creating indexes CONCURRENTLY, as that can't happen inside a
transaction.

  from psycopack import psycopack, get_db_connection

  with get_db_connection("postgresql://user:password@host:port/db") as conn:
      with conn.cursor() as cur:
          cur.execute("ABORT;")
          psycopack = Psycopack(
              batch_size=50_000,
              conn=conn,
              cur=cur,
              schema="public",
              table="to_psycopack",
              # Set this argument and Psycopack will do the conversion
              # automatically.
              convert_pk_to_bigint=True,
          )
          psycopack.sync_schemas()

In any case, Psycopack handles its own transactions and does not need to
rely on arbitrary library designs.

- [1] https://github.com/psycopg/psycopg2/issues/941#issuecomment-864025101
- [2] https://github.com/psycopg/psycopg2/issues/1305#issuecomment-866712961