Allow empty table (defaults to false)
Prior to this change, running psycopack on an empty table would cause the
TableIsEmpty exception to be raised.

This change adds an option to allow empty tables to be handled without
raising an exception. This enables psycopack to be used (e.g. for schema
changes) irrespective of whether the table is empty or not.