Add "logged" parameter to get_cursor
Prior to this change, get_cursor always returned a LoggedCursor, which logs
each SQL query executed at the debug level. That's helpful while developing
this library, but isn't so useful for applications using the library.

This change adds a "logged" boolean parameter to get_cursor, which defaults
to False. The change also updates the test suite to supply logged=True,
preserving the previous behaviour for tests.