Set up backfill log with separate positive and negative ranges
Prior to this change, if the table being repacked used negative pk values,
the backfill would cover the entire range from min(pk) to 0 (as well as the
positive values). When negative pk values have been used after positive
values are nearing exhaustion, it's likely that most of the negative range
is empty and therefore does not need to be backfilled.

This change checks for min(pk) and max(pk) separately for positive values
and negative values, to avoid backfilling empty negative ranges.