Fix queries to consider role membership as well as table ownership
Prior to this change, the introspections that looked to see whether the
current user owned a particular table only considered direct ownership.
But in the case of a table owned by a role, any user that has that role
also effectively owns the table, and has permission to operate on the table
as if they were the direct owner.

This change updates the get_referring_fks and is_table_owner introspection
functions to consider both direct ownership and also indirect ownership
via role membership.