Fix handling of PK sequence value
Prior to this change, the current value of the table's PK sequence
wasn't transferred to the repacked table's PK sequence. Without that,
new rows would be inserted into the repacked table starting with a PK
value of 1 (which would likely conflict with existing rows).

This change adds a command during swap to transfer the sequence value
from one table to the other. There is also special handling for the edge
case of a bigint conversion where the PK sequence has been set to negative
(presumably after positive PK values were exhausted); in that case, the
sequence is reset to the first positive bigint value not within the integer
range, which is 2**31. Note that this edge case won't be handled correctly
during a revert_swap(); addressing that is left as a TODO for later.