Disallow deferrable unique constraints
Prior to this change, if Psycopack was used on a table with a deferrable
unique constraint, then after the "sync_schemas" stage, any updates to the original
table would cause the trigger that updates the psycopack table to fail with
this error:
    ON CONFLICT does not support deferrable unique constraints/exclusion
     constraints as arbiters

That error results from the use of "ON CONFLICT DO NOTHING" in the copy
function used by the trigger.

This change removes test coverage of deferrable unique constraints and adds
a check to the pre_validate stage to reject a table that has a deferrable
unique constraint.