Update Registry table to record strategy update
This change adds four new columns to the Registry table:

- sync_strategy
- change_log
- change_log_function
- change_log_trigger

The change also handles updating existing Registry tables for users of
Psycopack who already started processing their tables so that they can
also get the update for these fields.

This is part of a larger work to allow Psycopack to have a new data and
schema synchronisation strategy that uses a change log table.