Add POST_SYNC_UPDATE tracker stage
The new synchronisation strategy requires a new step after synchronising
the schema so that all the changes that happened in the original table
since the Psycopack process began can be processed through.

This commit only adds the changes required by the tracker, leaving the
implementation of this stage to a subsequent commit.